backend/requirements.txt
aiosmtplib==2.0.2
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
dramatiq==1.17.0
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
fastapi-mail==1.4.1
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
pika==1.3.2
prometheus_client==0.20.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
setuptools==71.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
uvicorn==0.30.3
watchdog==4.0.1
watchdog_gevent==0.2.0
watchfiles==0.22.0
websockets==12.0
zope.event==5.0
zope.interface==6.4.post2
